ProjectPythonAI/requirements.txt
torch==2.0.0
torchvision==0.15.1
timm==0.9.0
transformers==4.40.0
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
Pillow==10.0.0
numpy==1.26.0
pandas==2.2.0
scikit-learn==1.4.0
matplotlib==3.8.0
tqdm==4.66.0
seaborn==0.13.0
kaggle==1.6.0